[BEAM-10849] Refactor to fix DataflowRunnerTest mocks.
